fix(ai): stabilize tool registry scaffold
